Add back inputbot global key bind listeners
TODO: We are currently doubly listening to media keys, the global hotkey library supports key press and release which we need to handle a press and hold correctly, however it doesn't work for full screened apps. We also use inputbot which works for full screened apps, but doesn't support key release events. Hence the redundancy for now...